requirements.txt
openai==1.75.0
numpy==1.26.4
tqdm==4.67.1
python-dotenv==1.0.1
tenacity==9.1.2
aiolimiter==1.2.1
typer[all]==0.15.2
tiktoken==0.7.0